backend/requirements.txt
fastapi
uvicorn
opencv-python
mediapipe
python-multipart
numpy
replicate
python-dotenv
